feature(connections): clean up url options
